[nrf temphack] REVERTME: carry undesired flash map change
The file flash_map_extended.c has some problems in its downstream NCS
history. Here are the relevant pieces (downstream on the left,
upstream on the right):

$ git log --oneline --graph 5f81b12~..d752124 \
                       boot/zephyr/flash_map_extended.c
*   d752124 [nrf mergeup] merge upstream up to commit f9be7a9
|\
| * 419a475 zephyr: allow dynamic numeration of flash_areas
| * 73bb71b CONFIG_FLASH_BASE_ADDRESS is defined only for memory-mapped Flash
| * 9f1933d boot: zephyr: migrate to new log subystem
| * 40c28e3 Add support for intel_s1000 SoC which is based on Xtensa arch
* | 549bdd5 boot: zephyr: migrate to new log subystem
* | 06ef990 Add support for intel_s1000 SoC which is based on Xtensa arch
|/
* 5f81b12 zephyr: fix flash_area_read_is_empty

In particular, it looks like "Add support for intel_s1000 SoC which is
based on Xtensa arch" getting cherry-picked downstream (as 06ef990)
before the mergeup (d752124) is causing a discrepancy with upstream,
because it made git merge not apply changes from 73bb71b to the
downstream version of this file.

Note there is a pretty large commit range which cherry-picked
downstream; this is just the part that is causing an issue with the
final contents of a file. There doesn't appear to be anything
malicious going on and the patch itself is harmless, but the
downstream cherry-picks should probably not have happened this way.

To rewrite downstream history in a way that gets to zero diff with a
previous snapshot, carry the patch forward for now.

Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>