[nrf toup] cmake: make update.hex without trailer data
This is needed for serial recovery to work using hex files.
Prior to this the update.hex got TLV data at the end of the
partition, which caused many blank pages to be included,
which made it hard to use in a serial recovery scheme.

Instead, make update.hex without TLV data at the end,
and provide a new file test_update.hex which contains
the TLV data, and can be directly flashed to test the
upgrade procedure.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>