[nrf fromtree] doc: fix link to external page in PORTING
Remove a footnote that is not generating a proper link and add an inline
link to the mbed TLS referece for platform.h. This also fixes a warning
when running through recommonmark==0.6.0 because it is unable to parse
the old syntax.

PR: JuulLabs-OSS/mcuboot#822

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>
(cherry picked from commit 82828b5cee28499157c340454e8cfa7a9beecd98)