[nrf noup] boot: zephyr: Add multi-image configuration to pm.yml
Adds definitions for `mcuboot_primary_1` and `mcuboot_secondary_1` to
the `pm.yml` file of MCUBoot. Which makes it configurable to enable and
disable these partitions by using the MCUBoot Kconfig.

Also it puts them in their respective region depending on the
configuration you have set for MCUBoot if it is going to use external
flash or mocked flash in RAM.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>