[nrf noup] kconfig: reduce downstream patch
squash! [nrf noup] treewide: add NCS partition manager support

Source the file 'Kconfig.mcuboot' from sdk-nrf to reduce the amount
of downstream patches in boot/zephyr/Kconfig.

NOTE TO UPMERGE AUTHOR:
This commit removes all but one downstream patches to the file
boot/zephyr/Kconfig. The only remaining patch should be the
sourcing of the Kconfig file in the mcuboot module in the nrf repo.

Hence there should be no modification to this file by any of the
[nrf noup] commits (except one)

noup commits with modifications that should be discarded:
[nrf noup] treewide: add NCS partition manager support
        4eba7803b69e0cb72f5d8d35584b1673df008851
        ! With the exception of the one line added by squashing this
        ! commit.

[nrf noup] kconfig: use cc310 for ecdsa by default
        a2422d91c88aa3415d471e6b213ac069d5ccf178

[nrf noup] boot: bootutil: Add shared crypto for ECDSA and SHA
        8ac5bd1f44a55968149721023c08ea83fef5790b

[nrf noup] boot: add 'minimal' configuration files
        2b23edc81b74f33fd2df0ece588d9ec73e5697b9

[nrf noup] boot: zephyr: add support for Thingy:91
        39fe200993fbf133e5023995d1bad3a140514540

[nrf noup] do_boot: clean peripherals state before boot
        9a3853b344f378cc6f03495cd8631815643506ed

Ref: NCSDK-11177

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>