[nrf fromtree] boot: zephyr: remove thingy53_nrf5340_cpuapp.conf board configuration
Removed the board configuration for Thingy:53 Application Core as it
contains references to the Kconfig modules that are not available in
the upstream Zephyr. The current configuration is set up to work
in the nRF Connect SDK environment and should be moved there.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>
(cherry picked from commit b9e47b76ed43209f88c02b945b0ffa9b022e993c)