[nrf fromtree] boot: zephyr: remove thingy53_nrf5340_cpuapp.conf board configuration
Removed the board configuration for Thingy:53 Application Core as it
contains references to the Kconfig modules that are not available in
the upstream Zephyr. The current configuration is set up to work
in the nRF Connect SDK environment and should be moved there.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>
(cherry picked from commit a1de4e55ef6ecfcba8fffb98f8cc08f19f05ce6c)