[nrf noup] boot: zephyr: move thingy53_nrf5340_cpuapp.conf downstream
Moved the board configuration for Thingy:53 Application Core to the
nRF Connect SDK MCUboot downstream repository. The configuration file
contains references to the Kconfig modules that are only available in
the nRF Connect SDK. The current configuration is set up to work in the
nRF Connect SDK environment and cannot be used upstream.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>