[nrf noup] treewide: add NCS partition manager support
fixup! [nrf noup] treewide: add NCS partition manager support

The comment correctly states that we should skip the check for image
1, but the code skips validation for all images except image 0.

Today it is only supported for two images, 0 and 1, but in the future
it is expected that there might be an image 2, so we refactor this
code to make it future-proof.

Also, the comment was confusing and misleading so it has been
re-written.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>