[nrf fromtree] zephyr: serial_recovery: Fix broken CDC device selection
When using CDC only and CONFIG_SERIAL is not enabled in Zephyr,
but there is chosen zephyr,mcumgr-uart, the chosen
takes precedence over CDC and directs MCUmgr to that uart.
This causes two issues: first the CDC is ignored if chosen exists,
and makes build fail because drivers are not built.

(cherry picked from commit 5397c13d240890f9866ef8b921cef67bd258fc25)
Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>