[nrf noup] bootutil: loader.c: Fix network core update external flash
Using sequential updates on the nRF53 with external flash would cause
network core updates to fail. This is because the update would be
located in the external flash which is network core cannot access. This
fixes the issue by copying the data from the external flash into RAM of
the application core which can be read by the network core.

This commit also contains some minor renaming to align with the naming
scheme used in the `loader.c` file for flash areas.

Ref: NCSDK-21379

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>