[nrf noup] sysflash: pm_sysflash: Fix incorrect define exclude for NSIB
A mismatch in defines was made for `_image_1_primary_slot_id` resulting
in some configurations not working. This fixes that the linker variable
is exposed for mcuboot so that it knows which slot is running an which
slot a bootloader upgrade is to be put into.

Ref. NCSDK-19223

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>
(cherry picked from commit 1f38eb867091b3dc8142d68f09545a8d9e3ed95e)