[nrf noup] zephyr: Boot even if EXT_ABI is not provided
This removes the `return;` to ensure that the application is booted even
if EXT_ABI is not provided to the application because it does not
include `FW_INFO`.

Added a bit more description to the error messages when FW_INFO is not
found and EXT_ABI is not able to be provided to the next image.

Ref. NCSDK-24132

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>
(cherry picked from commit d07555bf0439aeec7c15609c46a85a06fa9be807)
(cherry picked from commit 0f20d8ef9909c6c508024d9ea04bcfaaf4929dc5)