[nrf noup] boot: zephyr: Boot even if EXT_ABI is not provided
This removes the `return;` to ensure that the application is booted even
if EXT_ABI is not provided to the application because it does not
include `FW_INFO`.

Added a bit more description to the error messages when FW_INFO is not
found and EXT_ABI is not able to be provided to the next image.

Ref. NCSDK-24132

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>
(cherry picked from commit b80046d57c768fd9297f7921b323396359c5c2ca)
(cherry picked from commit 01f17ebce077e2f736c28ae4fb4215081e5b21b3)
Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>