[nrf noup] boot/zephyr: fix fw_info search
By the upstream patch the vt get now the pointer to the copy
of the arm_vector instead of original.

This patch fixes address of the firmware which is to be taken by
the fw_info_find.

Signed-off-by: Andrzej Puzdrowski <andrzej.puzdrowski@nordicsemi.no>
(cherry picked from commit 592f636c54a3b4c702c8e90a0d791ab1f00c4082)