boot: bootutil: Fix last sector index computation for swap-offset
When using swap-offset, the index of the last sector of the primary slot
that have to be swapped is computed at the beginning of the upgrade
process. This computation was in fact returning the number of sector to
swap rather than the index of the last sectpr (so N+1 instead of N).
This was causing the upgrade to fail for large image, using all the
available sectors.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>
(cherry picked from commit 7330df7c783b0da75190459860257fc9f758fd50)
(cherry picked from commit 870acb4d81d6d950777aa60c129e424146c27df2)