[nrf fromtree] boot: bootutil: swap-offset: Fix image size check during validation
When checking the size of an image in bootutil_img_validate, the offset
to the end of the TLV area was used as the image size in all cases.
However, when using swap-offset, the upgrade image is written in the
secondary with an offset. This offset is not part of the image and must
therefore not be taken into account in the image size.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>
(cherry picked from commit 2e6051915762465030028c15d45374e398dd01ed)