[nrf fromtree] sim: Use largest image for all upgrade strategies
The simulator was testing the upgrade with the largest image possible
for all strategies, except for overwrite-only, swap-move and swap-offset
because some tests were failing when the maximum image size was used.

For overwrite-only, this was due to an incorrect trailer size
computation. This has been fixed by
88294be188c25e4b7d02cc9c55fabfc9286c83b7.

For swap-move and swap-offset, this was due to the simulator not taking
into account the padding needed by those strategies in the primary or
secondary slot, but also to incorrect computation of the maximum image
size in some cases by the MCUboot library. Both issues have been fixed
by the previous commits.

Since all those issues have been fixed, the simulator can now be
configured to test upgrade with the largest possible image for all
strategies.

Note that logic needed to generate image of a given image is kept even
if not useful anymore at the moment, since that might be needed when
test will be added to ensure proper behavior when images of different
sizes are used.

Signed-off-by: Thomas Altenbach <thomas.altenbach@legrand.com>

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>
(cherry picked from commit f7394e395d189807dd205f56c6671e0f28af8f9a)