[nrf noup] zephyr: boards: align nrf54h20 with iron
The nrf54h20 iron variant was removed, as nrf54h20 is now iron
by default. The overlay for it was not updated though, so that
is done with this commit.

This commit shall be removed once upstream and downstream are
aligned on not using the iron variant of the nrf54h20.

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>