[nrf noup] boot: zephyr: Surpress bootutil_decode_sig warning
nrf-squash! [nrf noup] boot: Add shared crypto for ECDSA and SHA

If ASN1 was not enabled (in case of PSA crypto) a "defined but
not used" warning was issued.
This changes could not be placed in upstream as the line was
modified by commit e3d60913f17f2a520ec502c4af4ed19a329f1ab1

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>