[nrf fromlist] bootutil: Add MCUBOOT_USE_TLV_ALLOW_LIST
The ALLOW_ROGUE_TLVS is used to turn off TLV filtering in
code, basically to prevent processing TLVs that MCUboot is not
compiled to serve anyway.
The commit replaces identifier ALLOW_ROGUE_TLVS with
MCUBOOT_USE_TLV_ALLOW_LIST and reverse the logic around it, as
it now means opposite to the original.
This gives the feature an identifier in style of the
mcuboot_config.h defined identifiers.

Upstream PR #: 2410

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>