[nrf fromlist] boot: zephyr: Fix IO-based entrance method
If a device uses RESETINFO, than there are some bits set in the
resetinfo, even for reboots that should allow to interpret and
enter the device recovery.
That way it is possible to recover a device through serial recovery if
the main application resets due to i.e. watchdog.

Upstream PR #: 2468

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>