[nrf noup] nrf_cleanup: temporary GRTC cleanup with no counter reset
nrf-squash! [nrf noup] nrf_cleanup: nRF54h: fix missing peripheral cleanup

The GRTC should not reset its value and stop running when switching for
MCUBoot to application. Also in general, this should not be done in case
of reset - so the nrfx_grtc_unit function should not reset the counter
value. This will be fixed in nrfx (NRFX-8487), but the solution
is needed quicker. Hence, a temporary uninit function for GRTC was
added in nrf_cleanup, fixing this issue for MCUBoot.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>