[nrf noup] bootutil: Provide support for embedded AES keys
Commit provides support for MCUBOOT_EMBEDDED_ENC_KEY config
option, that allows to compile code with embedded key.
When this option is enabled, compilation requires
definition of boot_take_enc_key function to be provided by user;
prototype for the function is provided.

The boot_take_enc_key function is supposed to provide encryption
AES key to be used for image encryption and decryption.

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>