[nrf noup] Make fprotect area span only MCUBoot
nrf-squash! [nrf noup] treewide: add NCS partition manager support

This commit changes the fprotect area size definition.
This is necessary, as if TFM is used, tfm_storage is placed
between mcuboot and the application. The old implementation
led to the protected fprotect area including both
MCUBoot and tfm_storage. This was unnecessary and
in case of nRF54L15 (where the maximum FPROTECT area is small)
not being able to use FPROTECT in case TF-M was used.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>